# Conflicts:
#	Explorer/Assets/DCL/AvatarRendering/AvatarShape/Assets/AvatarDebugView.prefab
#	Explorer/Assets/Scripts/ECS/Profiling/ProfilerView.prefab
#	Explorer/Assets/Scripts/ECS/Profiling/ProfilingView.cs

+++ Explorer/Assets/Scripts/ECS/Profiling/IProfilingProvider.cs
@@ -6,7 +6,7 @@ namespace ECS.Profiling
 
         double GetAverageFrameTimeValueInNS();
 
-        int GetHiccupCountInBuffer();
+        ulong GetHiccupCountInBuffer();
 
         void CheckHiccup();
     }
